fix: split normal and shell compatibility functions into two separate files
- this is due to the fact. that some resources, like Shell can't be imported in the preferences
- fix all imports of those compatibility functions
