+++ src/components/searchBox.ts
@@ -5,11 +5,11 @@ import GObject from '@girs/gobject-2.0';
 import Meta from '@girs/meta-16';
 import Shell from '@girs/shell-16';
 import St from '@girs/st-16';
-import { orientationCompatibility } from '@pano/utils/compatibility';
 import { ItemType } from '@pano/utils/db';
 import { registerGObjectClass, SignalRepresentationType, SignalsDefinition } from '@pano/utils/gjs';
 import { getPanoItemTypes, ICON_PACKS } from '@pano/utils/panoItemType';
 import { getCurrentExtensionSettings, gettext } from '@pano/utils/shell';
+import { orientationCompatibility } from '@pano/utils/shell_compatibility';
 
 export type SearchBoxSignalType =
   | 'search-text-changed'
